aligning with generalizing tp2
